requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.18.0
tensorflow>=2.10.0
keras>=2.10.0
Pillow>=9.0.0
matplotlib>=3.5.0
pandas>=1.3.0
